requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
neo4j==5.6.0
python-dotenv==1.0.0
pytz==2022.7.1
Werkzeug==2.2.3
